OPENDJ-1795 On Windows: "multiple SLF4J bindings" error while using stop-ds tool against a remote instance
Problem is down to scripts calling each other.
stop-ds.bat calls
  set SCRIPT_UTIL_CMD=set-full-server-environment-and-test-java
then goes to stopUsingProtocol label where it "calls" "_client-script.bat"
which then calls
  set SCRIPT_UTIL_CMD=set-full-environment

And the CLASSPATH variable ends up with bootstrap.jar (with server logger) and bootstrap-client.jar (with java.util.logging logger) in it.



On Windows batch, "call" executes a script in the current bash context. It seems to be the equivalent of bash's "source".
The fix is to empty the CLASSPATH before calling "_client-script.bat"

git-svn-id: https://svn.forgerock.org/opendj/trunk@12238 41b1ffd8-f28e-4786-ab96-9950f0a78031