CR-7022 OPENDJ-2013 Build without local-db-backend doc
This patch uses conditional text to make it possible
to exclude elements related to the local-db-backend
from the hand-written documentation.

This patch does not address local-db-backend related content
in the generated reference documentation.
For the most part, that should be a by-product of changes
to OpenDJ directory server code, though there is one place
in the opendj-server-legacy POM that I've left a comment
regarding the need to avoid trying to build a man page for dbtest.

git-svn-id: https://svn.forgerock.org/opendj/trunk@12315 41b1ffd8-f28e-4786-ab96-9950f0a78031