CR-6989 OPENDJ-1946 Doc how to reinitialize replication
This patch adds a procedure to reinitialize data for a base DN.

This procedure uses the pre-/post-external-init subcommands,
rather than rm -rf of the changelogDb and database files.
The approach with rm is perhaps quicker for test/PoC environments,
but looks like it calls for shutting the service down.

git-svn-id: https://svn.forgerock.org/opendj/trunk@12324 41b1ffd8-f28e-4786-ab96-9950f0a78031