CR-7072 OPENDJ-2060 Build config reference again
This patch adds a Maven plugin to build the configuration reference
alongside the other server docs.

This patch also makes some minor changes to comment out
links in the configuration reference that might break
and that are not strictly necessary.

Finally this patch refactors the OpenDJ doc plugin
where some methods are used across Mojos.

git-svn-id: https://svn.forgerock.org/opendj/trunk@12362 41b1ffd8-f28e-4786-ab96-9950f0a78031