OPENDJ-2108 CR-7212 Some indexes are evaluated as indexed but should be not-indexed.
Running a search query with debugsearchindex, should tell how the server uses indexes.
Some optimisations change the index type actually used to perform, so that even if the attribute or the type are not indexed,
we may revert to an indexed search but we report the original search attribute and index type.


git-svn-id: https://svn.forgerock.org/opendj/trunk@12488 41b1ffd8-f28e-4786-ab96-9950f0a78031