Code cleanup
*.java:
Used CollectionUtils.newArrayList() as much as possible.
Made use of LDAPAttribute constructors which accepts Strings. This avoids String to ByteString conversion in client code since the constructors handle it.

AttributeBuilder.java:
Added toAttributeList() to avoid the need for client code to create a List itself.

Attributes.java:
Added several createAsList() to avoid the need for client code to create a List itself.

git-svn-id: https://svn.forgerock.org/opendj/trunk@12642 41b1ffd8-f28e-4786-ab96-9950f0a78031