OPENDJ-2171 CR-7414 import-ldif in append mode should not reset index state
Regardless of options specified to import-ldif, all indexes are systematically set as not trusted before importing and reset to trusted once it's finished. 
On the other hand if the suffix is simply appended to, indexes not up to date with the data should not be set to trusted, since they may end up with partial data. 

Before import we should set all trusted indexes as candidates for updates, but leave not trusted indexes alone.
In order to do that, use index lists from Suffix, building the list of indexes to update taking into account the append flag: if it is specified only consider trusted indexes. 
Import code should only get index information from the Suffix. 

Rebuild index should still be based on the EntryContainer though, since we are managing indexes directly.


git-svn-id: https://svn.forgerock.org/opendj/trunk@12646 41b1ffd8-f28e-4786-ab96-9950f0a78031